+++ eric_sse/benchmark.py
@@ -1,6 +1,6 @@
 import time
 
-from eric_sse.entities import MessageQueueListener
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import MessageContract
 from eric_sse.prefabs import DataProcessingChannel
 from eric_sse import get_logger
