+++ eric_sse/entities.py
@@ -1,13 +1,14 @@
 import asyncio
 import traceback
 from abc import ABC, abstractmethod
-from threading import Lock
 from typing import AsyncIterable, Any
 
 import eric_sse
 from eric_sse.exception import InvalidListenerException, NoMessagesException, InvalidChannelException
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import MessageContract, Message
-from eric_sse.queue import Queue, AbstractMessageQueueRepository, InMemoryMessageQueueFactory
+from eric_sse.queue import Queue
+from eric_sse.repository import AbstractMessageQueueRepository, InMemoryMessageQueueRepository
 
 logger = eric_sse.get_logger()
 
@@ -15,47 +16,12 @@ MESSAGE_TYPE_CLOSED = '_eric_channel_closed'
 MESSAGE_TYPE_END_OF_STREAM = '_eric_channel_eof'
 MESSAGE_TYPE_INTERNAL_ERROR = '_eric_error'
 
-class MessageQueueListener(ABC):
-    """
-    Base class for listeners.
-
-    Optionally you can override on_message method if you need to inject code at message delivery time.
-    """
-
-    def __init__(self):
-        self.id: str = eric_sse.generate_uuid()
-        self.__is_running: bool = False
-
-    async def start(self) -> None:
-        self.start_sync()
-
-    def start_sync(self) -> None:
-        logger.debug(f"Starting listener {self.id}")
-        self.__is_running = True
-
-    async def is_running(self) -> bool:
-        return self.is_running_sync()
-
-    def is_running_sync(self) -> bool:
-        return self.__is_running
-
-    async def stop(self) -> None:
-        self.stop_sync()
-
-    def stop_sync(self) -> None:
-        logger.debug(f"Stopping listener {self.id}")
-        self.__is_running = False
-
-    def on_message(self, msg: MessageContract) -> None:
-        """Event handler. It executes when a message is delivered to client"""
-        pass
-
 
 class ConnectionManager:
-    def __init__(self, queues_factory: AbstractMessageQueueRepository):
+    def __init__(self, queues_repository: AbstractMessageQueueRepository):
         self.__listeners: dict[str: MessageQueueListener] = {}
         self.__queues: dict[str: Queue] = {}
-        self.__queues_factory = queues_factory
+        self.__queues_factory = queues_repository
 
 
     def add_listener(self) -> MessageQueueListener:
@@ -88,6 +54,8 @@ class ConnectionManager:
         except KeyError:
             raise InvalidListenerException
 
+    def get_listeners(self) -> dict[str, MessageQueueListener]:
+        return self.__listeners
 
 class AbstractChannel(ABC):
     """
@@ -100,24 +68,19 @@ class AbstractChannel(ABC):
 
     :param int stream_delay_seconds: Wait time in seconds between message delivery.
 
-    :param eric_sse.queue.AbstractMessageQueueRepository queues_factory:
+    :param eric_sse.repository.AbstractMessageQueueRepository queues_repository:
     """
     def __init__(
             self,
             stream_delay_seconds: int = 0,
-            queues_factory: AbstractMessageQueueRepository | None = None
+            queues_repository: AbstractMessageQueueRepository | None = None
     ):
         self.id: str = eric_sse.generate_uuid()
         self.stream_delay_seconds = stream_delay_seconds
 
-        self.__listeners: dict[str: MessageQueueListener] = {}
-        self.__queues: dict[str: Queue] = {}
-        self.__queues_factory = queues_factory if queues_factory else InMemoryMessageQueueFactory()
+        queues_repository = queues_repository if queues_repository else InMemoryMessageQueueRepository()
+        self.__connection_manager: ConnectionManager = ConnectionManager(queues_repository)
 
-        self.__connection_manager: ConnectionManager = ConnectionManager(self.__queues_factory)
-
-    def _set_queues_factory(self, queues_factory: AbstractMessageQueueRepository):
-        self.__queues_factory = queues_factory
 
     def add_listener(self) -> MessageQueueListener:
         """Add the default listener"""
@@ -158,7 +121,7 @@ class AbstractChannel(ABC):
 
     def broadcast(self, msg: MessageContract):
         """Enqueue a message to all listeners"""
-        for listener_id in self.__listeners.keys():
+        for listener_id in self.__connection_manager.get_listeners():
             self.dispatch(listener_id, msg=msg)
 
     def get_listener(self, listener_id: str) -> MessageQueueListener:
