+++ eric_sse/prefabs.py
@@ -2,10 +2,11 @@ from asyncio import as_completed as as_completed_future, get_running_loop
 from concurrent.futures import ThreadPoolExecutor, Executor
 from typing import Callable, AsyncIterable
 from eric_sse import get_logger
-from eric_sse.entities import AbstractChannel, MessageQueueListener
+from eric_sse.entities import AbstractChannel
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import SignedMessage, MessageContract
 from eric_sse.exception import NoMessagesException
-from eric_sse.queue import AbstractMessageQueueRepository
+from eric_sse.repository import AbstractMessageQueueRepository
 
 logger = get_logger()
 
@@ -19,16 +20,16 @@ class SSEChannel(AbstractChannel):
 
     :param int stream_delay_seconds:
     :param int retry_timeout_milliseconds:
-    :param eric_sse.queue.AbstractMessageQueueRepository queues_factory:
+    :param eric_sse.repository.AbstractMessageQueueRepository queues_repository:
     """
 
     def __init__(
             self,
             stream_delay_seconds: int = 0,
             retry_timeout_milliseconds: int = 5,
-            queues_factory: AbstractMessageQueueRepository = None
+            queues_repository: AbstractMessageQueueRepository = None
     ):
-        super().__init__(stream_delay_seconds=stream_delay_seconds, queues_factory=queues_factory)
+        super().__init__(stream_delay_seconds=stream_delay_seconds, queues_repository=queues_repository)
         self.retry_timeout_milliseconds = retry_timeout_milliseconds
 
         self.payload_adapter: (
