+++ eric_sse/queue.py
@@ -44,24 +44,6 @@ class InMemoryQueue(Queue):
     def delete(self) -> None:
         self.__messages = []
 
-class AbstractMessageQueueRepository(ABC):
-    """
-    Abstraction for queues creation
-
-    see :class:`eric_sse.entities.AbstractChannel`
-    """
-    @abstractmethod
-    def create(self) -> Queue:
-        ...
-
-
-class InMemoryMessageQueueFactory(AbstractMessageQueueRepository):
-    """
-    Default implementation used by :class:`eric_sse.entities.AbstractChannel`
-    """
-    def create(self) -> Queue:
-        return InMemoryQueue()
-
 
 class RepositoryError(Exception):
     """
