+++ examples/sockets/server.py
@@ -1,6 +1,6 @@
 from asyncio import run, CancelledError
 
-from eric_sse.entities import MessageQueueListener
+from eric_sse.listener import MessageQueueListener
 from eric_sse.message import Message
 from eric_sse.servers import SocketServer
 from examples.sockets import SOCKET_FILE_DESCRIPTOR_PATH
